Rename core to abstractions
