+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/DataAccess/MeteringPoints/Queries/MeteringPointByGsrnQueryHandler.cs
@@ -17,12 +17,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
 using Energinet.DataHub.MeteringPoints.Application.Queries;
-using Energinet.DataHub.MeteringPoints.Client.Core;
+using Energinet.DataHub.MeteringPoints.Client.Abstractions.Models;
 using MediatR;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.MeteringPoints.Queries
 {
-    public class MeteringPointByGsrnQueryHandler : IRequestHandler<MeteringPointByGsrnQuery, MeteringPointDTO?>
+    public class MeteringPointByGsrnQueryHandler : IRequestHandler<MeteringPointByGsrnQuery, MeteringPointDto?>
     {
         private readonly IDbConnectionFactory _connectionFactory;
 
@@ -31,7 +31,7 @@ namespace Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.MeteringPoi
             _connectionFactory = connectionFactory;
         }
 
-        public async Task<MeteringPointDTO?> Handle(MeteringPointByGsrnQuery request, CancellationToken cancellationToken)
+        public async Task<MeteringPointDto?> Handle(MeteringPointByGsrnQuery request, CancellationToken cancellationToken)
         {
             if (request == null) throw new ArgumentNullException(nameof(request));
 
@@ -89,7 +89,7 @@ namespace Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess.MeteringPoi
 
             var result = await _connectionFactory
                 .GetOpenConnection()
-                .QuerySingleOrDefaultAsync<MeteringPointDTO>(sql, new { request.GsrnNumber })
+                .QuerySingleOrDefaultAsync<MeteringPointDto>(sql, new { request.GsrnNumber })
                 .ConfigureAwait(false);
 
             return result;
